% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateLogRegCV.R
\name{aggregateLogRegCV}
\alias{aggregateLogRegCV}
\title{Aggregate the results from a logRegCV() run}
\usage{
aggregateLogRegCV(logRegCV.object, var)
}
\arguments{
\item{logRegCV.object}{A list object generated by logRegCV()}

\item{var}{A character string of what information to return. 'lambda.min': lambda values with the lowest error, 'lambda.1se':
lambda value one standard error higher than lambda.min, 'coef': regression coefficients, 'pred': probabilities from predictions
on the outer fold test sets}
}
\value{
Returns a vector/matrix of the selected result type
}
\description{
Combines the results from a list generated by logRegCV() into a matrix. Useful for making summary plots.
}
\examples{
## Extract lambda.min
aggregateLogRegCV(lrCV, 'lambda.min')

## Extract multiple result types
lapply(c('lambda.min','coef','pred'), function(i){ aggregateLogRegCV(lrCV, i) })
}
