% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateRandomForestCV.R
\name{aggregateRandomForestCV}
\alias{aggregateRandomForestCV}
\title{Aggregate the results from a randomForestCV() run}
\usage{
aggregateRandomForestCV(randomForestCV.object, var)
}
\arguments{
\item{randomForestCV.object}{A list object generated by randomForestCV()}

\item{var}{A character string of what information to return. 'importance': MeanDecreaseAccuracy, 'pred': probabilities
from predictions}
}
\value{
Returns a dataframe/matrix of the selected result type
}
\description{
Aggregate the results from a randomForestCV() run
}
\examples{
## Extract MeanDecreaseAccuracy
aggregateLogRegCV(rfCV, 'importance')

## Extract multiple result types
lapply(c('importance','pred'), function(i){ aggregateRandomForestCV(rfCV, i) })
}
