% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceClasses.R
\name{balanceClasses}
\alias{balanceClasses}
\title{Balancing of classes by up/down sampling}
\usage{
balanceClasses(df, colname.response, target.class, scale.ratio,
  method = "simple")
}
\arguments{
\item{df}{A dataframe of samples as rows and features as columns}

\item{colname.response}{The name of a column containing the response classes}

\item{target.class}{The target class for which the sample size will be adjusted}

\item{scale.ratio}{The amount of up/down-sampling (e.g. 2 is 2x up, 0.5 is 2x down)}

\item{method}{'simple': random samples are replicated when upsampling or random samples are removed when downsampling.
'SMOTE': synthesizes samples based on the feature values of a class (can only be used for upsampling)}
}
\value{
Returns a dataframe with balanced classes
}
\description{
Data has class imbalance if e.g. there are 100 BRCA deficient donors and 900 BRCA proficient donors. This 
function up/down-samples a desired class.
}
\examples{
balanceClasses(df, 'response', 'BRCA1', 1.5, 'SMOTE')
}
