% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceClassesMulti.R
\name{balanceClassesMulti}
\alias{balanceClassesMulti}
\title{Up/down sample a dataframe containing multiple classes}
\usage{
balanceClassesMulti(df, colname.response, balance.options)
}
\arguments{
\item{df}{A dataframe of observations as rows and features as columns}

\item{colname.response}{The name of a column containing the response classes}

\item{balance.options}{A list of vectors stating the options of the base function, balanceClasses(). Vectors should be in the 
format: c(target.class, scale.ratio, method)}
}
\value{
A dataframe of resampled classes according to the balance.options provided
}
\description{
Up/down sample a dataframe containing multiple classes
}
\examples{
balance.options = list(
   c('none', 1/3, 'simple'),
   c('BRCA1', 2, 'SMOTE')
)

balanceClassesMulti(df, 'response', balance.options)
}
