% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAuc.R
\name{calcAuc}
\alias{calcAuc}
\title{Calculate AUC}
\usage{
calcAuc(x, y)
}
\arguments{
\item{x}{A numeric/integer vector of discrete x values}

\item{y}{A numeric/integer vector of discrete y values}
}
\value{
Returns the area under the curve
}
\description{
Calculates the area under a curve (AUC) by integration via the trapezoidal rule, where discrete (x,y)
coordinates are given.
}
\examples{
calcAuc(1:10, runif(10))
}
