% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPerf.R
\name{calcPerf}
\alias{calcPerf}
\title{Calculate classifier performance metrics}
\usage{
calcPerf(confusion, metrics, melt = F, add.start.end.values = T)
}
\arguments{
\item{confusion}{A numeric vector or matrix or data frame. These should contain a confusion matrix for one cutoff or 
confusion matrices for multiple cutoffs.}

\item{metrics}{A character vector of the desired statistics. Multiple metrics can be specified.}

\item{melt}{If melt = FALSE, output in wide format. If melt = TRUE, output in long format}
}
\value{
A numeric vector or matrix of the selected performance metrics
}
\description{
Calculate classifier performance metrics
}
\examples{
calcPerf(confusion, c('tpr','tnr'))
}
