% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPerfCompound.R
\name{calcPerfCompound}
\alias{calcPerfCompound}
\title{Calculate compound performance metrics (e.g. roc)}
\usage{
calcPerfCompound(
  confusion,
  compound.metric,
  avg.method = NULL,
  metric.names.as.x.y = F
)
}
\arguments{
\item{confusion}{A matrix or data frame with each row being a confusion matrix. The output of 
confusionMatrix(probs.predicted, logicals.expected, cutoff = 'all')}

\item{compound.metric}{The name of the compound metric}

\item{avg.method}{How to average performance metrics in multiclass classification. Can be NULL, 
'weighted' or 'macro'}

\item{metric.names.as.x.y}{If TRUE, will convert colnames: 'tpr' and 'ppv', to: 'x' and 'y'}
}
\value{
A matrix of the two performance metrics that the compound metric is composed of
}
\description{
Calculate compound performance metrics (e.g. roc)
}
