% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusionMatrix.R
\name{calcTfpn}
\alias{calcTfpn}
\title{Calculate true/false positive/negative counts}
\usage{
calcTfpn(predicted = NULL, actual = NULL, cutoff = 0.5)
}
\arguments{
\item{predicted}{A numeric vector of prediction probabilities, or a logical
vector}

\item{actual}{A logical vector of the actual classes}

\item{cutoff}{A single classification cutoff from 0 to 1. Only required when
`predicted` are probabilities}
}
\value{
A integer vector of the true/false positive/negative counts
}
\description{
Calculate true/false positive/negative counts
}
