% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierPerf.R
\name{classifierPerf}
\alias{classifierPerf}
\title{Calculate classifier performance metrics}
\usage{
classifierPerf(confusion, metrics, melt = F)
}
\arguments{
\item{confusion}{A numeric vector or matrix or data frame. These should contain a confusion matrix for one cutoff or 
confusion matrices for multiple cutoffs.}

\item{metrics}{A character vector of the desired statistics. Multiple metrics can be specified.}

\item{melt}{If melt = FALSE, output in wide format. If melt = TRUE, output in long format}
}
\value{
A numeric vector; or matrix or data frame
}
\description{
Calculate classifier performance metrics
}
\examples{
classifierPerf(confusion, c('tpr','tnr'))
}
