% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusionMatrix.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{Calculate confusion matrix}
\usage{
confusionMatrix(probs.predicted, logicals.expected, cutoff = 0.5,
  as.vector = F)
}
\arguments{
\item{probs.predicted}{A numeric vector of prediction probabilities.}

\item{logicals.expected}{A character vector of response labels. If provided, dots will be coloured by these labels.}

\item{cutoff}{If provided, a horizontal line will be drawn at the specified value to indicate the chosen cutoff.}

\item{as.vector}{Return vector or matrix?}
}
\value{
A vector or matrix
}
\description{
Calculate confusion matrix
}
