% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusionMatrix.R
\name{confusionMatrixMC}
\alias{confusionMatrixMC}
\title{Create confusion matrix(-ces) multi-class}
\usage{
confusionMatrixMC(predicted, expected, cutoff = "all", neg.response = NULL)
}
\arguments{
\item{predicted}{A matrix or data frame containing the prediction probabilities of each class in each column.}

\item{expected}{A character vector of response labels.}

\item{cutoff}{A single or vector of classification cutoffs ranging from 0 to 1}

\item{neg.response}{The names of the negative response(s). These classes will be excluded from the calculations.}
}
\value{
A vector or matrix
}
\description{
DEPRECATED; DO NOT USE
}
