% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2BinaryResponse.R
\name{convert2BinaryResponse}
\alias{convert2BinaryResponse}
\title{Convert multiclass response to binary response}
\usage{
convert2BinaryResponse(chr, p.classes = NULL, p.return = NULL,
  n.classes = NULL, n.return = NULL, as.factor = TRUE)
}
\arguments{
\item{chr}{A character vector or factor}

\item{p.classes}{Name of positive classes to convert}

\item{p.return}{Name of positive class to return}

\item{n.classes}{Name of negative classes to convert}

\item{n.return}{Name of negative class to return}

\item{as.factor}{Option to return a factor instead of character vector}
}
\value{
Returns a character vector or factor with only two classes
}
\description{
Convert multiclass response to binary response
}
\details{
Classes here refer to the type of responses; not to be confused with R classes.

If only one class is specified for Pclasses and no name is specified for Preturn, Pclasses is returned. Preturn can be
specified in case one would like to change the name of the class. This is also the same for Nclasses/Nreturn.
}
\examples{
## Convert c('BRCA1','BRCA2') to 'BRCA'. Leave 'none'  as is.
convert2BinaryResponse(c('BRCA1', 'BRCA2', 'none'), c('BRCA1','BRCA2'), 'BRCA', 'none', 'none')
}
