% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCvTrainTestSets.R
\name{createCvTrainTestSets}
\alias{createCvTrainTestSets}
\title{Create train and test sets for k-fold cross validation}
\usage{
createCvTrainTestSets(df, k = 10, stratify.by.col = NULL, seed = NULL)
}
\arguments{
\item{df}{A dataframe of observations as rows and features as columns}

\item{k}{Number of folds to split the data into}

\item{stratify.by.col}{(Optional) The name of a column containing the response classes by which to stratify the data.
Stratification ensures that all classes are present in every train/test set.}
}
\value{
A list of length(k), containing a sublist of train and test sets
}
\description{
Shuffles observations and splits them into k-fold train/test sets.
}
\examples{
createCvTrainTestSets(df, k=10, stratify.by.col = 'response')
}
