% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusionMatrix.R
\name{getTfpn}
\alias{getTfpn}
\title{Calculate true/false positive/negative counts}
\usage{
getTfpn(probs.predicted, logicals.expected, cutoff)
}
\arguments{
\item{probs.predicted}{A numeric vector of prediction probabilities.}

\item{logicals.expected}{A character vector of response labels.}

\item{cutoff}{A single classification cutoff from 0 to 1}
}
\value{
A integer vector of the true/false positive/negative counts
}
\description{
Calculate true/false positive/negative counts
}
