% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ln.R
\name{ln}
\alias{ln}
\title{Natural logarithm}
\usage{
ln(x)
}
\arguments{
\item{x}{A numeric/integer vector}
}
\value{
Returns an ln() transformed numeric/integer vector
}
\description{
Natural logarithm
}
\examples{
ln(2)
}
