% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticRegression.R
\name{logisticRegression}
\alias{logisticRegression}
\title{Logistic regression equation}
\usage{
logisticRegression(featureValues, intercept, coefs)
}
\arguments{
\item{featureValues}{A numeric/integer vector containing values for each feature}

\item{intercept}{Intercept}

\item{coefs}{Coefficients for each feature}
}
\value{
Returns a value between 0 and 1 giving the probability of the positive outcome
}
\description{
Predict the probability of a binary outcome for an observation given the feature values, intercept and coefficients
}
