% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRegCV.R
\name{logRegCV}
\alias{logRegCV}
\title{Logistic regression nested k-fold cross-validation}
\usage{
logRegCV(df, colname.response, positive.response, k = 10, stratify = T, ...)
}
\arguments{
\item{df}{A dataframe of observations as rows and features as columns. \strong{Important:} input matrix/dataframe should
include the response vector as a column.}

\item{colname.response}{The name of a column containing the response classes.}

\item{k}{The number of outer cross-validation folds. Also the number of inner cross-validation folds (i.e. fitting the 
regression with \pkg{glmnet::cv.glmnet()} ).}

\item{stratify}{Stratification ensures that all classes are present in every train/test set when splitting the
data into k-folds.}

\item{...}{Other arguments that can be passed to \pkg{glmnet::cv.glmnet()} and \pkg{logRegTrainAndTest}.}
}
\value{
Returns a list of length(k), with sublists containing (1) a cv.glmnet object from the training on the outer
fold train set, and (2) a dataframe of probabilities from the prediction on the outer fold test set.
}
\description{
Performs nested k-fold cross-validation to assess the performance of a logistic regression model.
\strong{Important:} input matrix/dataframe should include the response vector as a column. Inner folds: uses
\pkg{glmnet::cv.glmnet()} fit a logistic regression to the outer fold training set (i.e. determine the optimal lambda and
subsequently the regression coefficients). Outer folds: uses the model created from the inner fold to predict on the outer
fold test set.
}
\examples{
logRegCV(df, colname.response = 'response', positive.response = 'BRCA', 
         k = 10, stratify= T, standardize = T, balance = 'up')

}
