% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneHotEncode.R
\name{oneHotEncode}
\alias{oneHotEncode}
\title{Converts a vector to a matrix of labels}
\usage{
oneHotEncode(labels, sample.names = NULL, as.integer = F)
}
\arguments{
\item{labels}{A vector of labels}

\item{sample.names}{Optional sample names}
}
\value{
A logical matrix
}
\description{
Converts a vector to a matrix of labels
}
