% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceAsDf.R
\name{performanceAsDf}
\alias{performanceAsDf}
\title{Calculate binary classification statistics}
\usage{
performanceAsDf(probs.predicted, logicals.expected, metrics = c("tpr", "tnr"),
  melt = F)
}
\arguments{
\item{probs.predicted}{A numeric vector of prediction probabilities.}

\item{logicals.expected}{A logical vector/factor stating the expected outcome of the observation.}

\item{metrics}{Default: c('tpr','tnr'). A character vector of the desired classification statistics. See documentation for
\pkg{ROCR:performance()} for more information of the available statistics.}

\item{melt}{If melt = FALSE, output in wide format. If melt = TRUE, output in long format}
}
\value{
Returns a long format dataframe of the desired statistics.
}
\description{
Uses \pkg{ROCR::prediction()} and \pkg{ROCR::performance()} to calculate classification statistics and outputs
a long or wide format dataframe suitable for plotting with \pkg{ggplot2}.
}
