% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_featureImportance.R
\name{plot_featureImportance}
\alias{plot_featureImportance}
\title{Plot feature importance}
\usage{
plot_featureImportance(df, title = NULL, sort = T, topn = F,
  rm.features.all.zero = F, final.model.imp = NULL, export.features = F)
}
\arguments{
\item{title}{User specified plot title. If unspecified, title defaults to 'Feature importance'}

\item{sort}{Whether to sort the features by decreasing importance}

\item{topn}{Whether to only show the top n ranked features. Depends on sort.}

\item{rm.features.all.zero}{Remove feature columns which contain all zeroes}

\item{final.model.imp}{If a vector is provided, red dashes will be added to the plot according to vector values. Used
for overlaying the feature importance of the final model over feature importance as determined by cross-validation}

\item{export.features}{If TRUE, only returns a vector containing the feature names}

\item{df.importance}{A data.frame containing a column of feature names and a column of MeanDecreaseAccuracy values}
}
\value{
Returns a ggplot2 scatter plot
}
\description{
Plot feature importance
}
