% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_logRegCoefs.R
\name{plot_logRegCoefs}
\alias{plot_logRegCoefs}
\title{Plot logistic regression coefficients}
\usage{
plot_logRegCoefs(data, title = NULL)
}
\arguments{
\item{title}{User specified plot title. If unspecified, the plot title defaults to 'Feature weights'.}

\item{df}{A dataframe/matrix with rows as features and columns as CV runs. Number of cols can be 1, to facilitate plotting
of coefficients from a final model}
}
\value{
Returns a ggplot2 boxplot
}
\description{
Plot logistic regression coefficients
}
