% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MeanDecreaseAccuracy.R
\name{plot_MeanDecreaseAccuracy}
\alias{plot_MeanDecreaseAccuracy}
\title{Plot MeanDecreaseAccuracy}
\usage{
plot_MeanDecreaseAccuracy(df, title = NULL, sort = T, topn = F)
}
\arguments{
\item{title}{User specified plot title. If unspecified, title defaults to 'Feature importance'}

\item{sort}{Whether to sort the features by decreasing importance}

\item{topn}{Whether to only show the top n ranked features. Depends on sort.}

\item{df.importance}{A data.frame containing a column of feature names and a column of MeanDecreaseAccuracy values}
}
\value{
Returns a ggplot2 scatter plot
}
\description{
Plot MeanDecreaseAccuracy
}
