% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_performanceRates.R
\name{plot_performanceRates}
\alias{plot_performanceRates}
\title{Line plot of binary classification statistics}
\usage{
plot_performanceRates(probs.predicted, logicals.expected, metrics = c("tpr",
  "tnr"), title = NULL, show.intersection = F, intersection.only = F)
}
\arguments{
\item{probs.predicted}{A numeric vector of prediction probabilities.}

\item{logicals.expected}{A logical vector/factor stating the expected outcome of each prediction.}

\item{metrics}{A character vector of the desired statistics. See documentation for \pkg{ROCR:performance()} for more
information of the available statistics. Multiple metrics can be specified; however, usually only two are used for
binary classification statistics plots.}

\item{title}{User specified plot title. If unspecified, the plot title defaults to the names of the statistics chosen.}

\item{show.intersection}{Show the intersection of 2 performance metrics.}

\item{intersection.only}{Only return the intersection value (cutoff value)}
}
\value{
Returns a ggplot2 line plot
}
\description{
Plots the rates of binary classification statistics (e.g. true positive rate, true negative rate) as a line plot.
}
