% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PR.R
\name{plot_PR}
\alias{plot_PR}
\title{Plot precision recall}
\usage{
plot_PR(probs.predicted, logicals.expected, title = NULL, show.auc = T,
  auc.only = F)
}
\arguments{
\item{probs.predicted}{A numeric vector of prediction probabilities.}

\item{logicals.expected}{A logical vector/factor stating the expected outcome of each prediction.}

\item{title}{User specified plot title. If unspecified, the plot title defaults to 'PR'.}

\item{show.auc}{Shows the area under the under value.}

\item{auc.only}{If TRUE, only returns the AUC value}
}
\value{
Returns a ggplot2 line plot
}
\description{
Plot precision recall
}
