% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ROC.R
\name{plot_ROC}
\alias{plot_ROC}
\title{Plot ROC curve}
\usage{
plot_ROC(probs_predicted, logicals_expected, title = NULL, showAUC = T)
}
\arguments{
\item{probs_predicted}{A numeric vector of prediction probabilities.}

\item{title}{User specified plot title. If unspecified, the plot title defaults to 'ROC'.}

\item{showAUC}{Default: TRUE. Shows the AUC on the plot.}

\item{logicals.expected}{A logical vector/factor stating the expected outcome of each prediction.}
}
\value{
Returns a ggplot2 line plot
}
\description{
Plot ROC curve
}
