% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sortedPredictedProbs.R
\name{plot_sortedPredictedProbs}
\alias{plot_sortedPredictedProbs}
\title{Plot sorted prediction probabilities}
\usage{
plot_sortedPredictedProbs(probs.predicted = NULL, annotation = NULL,
  cutoff = NULL, title = NULL, decreasing = F)
}
\arguments{
\item{probs.predicted}{A numeric vector of prediction probabilities.}

\item{annotation}{A character vector of response labels. If provided, dots will be coloured by these labels.}

\item{cutoff}{If provided, a horizontal line will be drawn at the specified value to indicate the chosen cutoff.}

\item{title}{User specified plot title.}

\item{decreasing}{Sort by decreasing?}
}
\value{
A ggplot2 scatter plot
}
\description{
Plot sorted prediction probabilities
}
