% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotCoxHazardRatios}
\alias{plotCoxHazardRatios}
\title{Plot Hazard Ratios from Cox Model}
\usage{
plotCoxHazardRatios(cox_fit)
}
\arguments{
\item{cox_fit}{A Cox proportional hazards model object.}
}
\description{
This function plots the hazard ratios from a Cox proportional hazards model.
}
\examples{
data <- preprocess_data(generate_employment_data())
cox_fit <- run_cox_model(data, covariates = c("Age", "Gender"))
plot_cox_hazard_ratios(cox_fit)
}
