% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotKmCurve}
\alias{plotKmCurve}
\title{Plot Kaplan-Meier Survival Curves}
\usage{
plotKmCurve(km_fit, data)
}
\arguments{
\item{km_fit}{A Kaplan-Meier survival object.}

\item{data}{The original data frame used to fit km_fit.}
}
\description{
Plot Kaplan-Meier Survival Curves
}
