% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPerf.R
\name{plotPerf}
\alias{plotPerf}
\title{Line plot of classification statistics}
\usage{
plotPerf(
  confusion,
  metrics = c("tpr", "tnr"),
  avg.method = NULL,
  title = NULL,
  show.intersection = F,
  intersection.only = F,
  legend.in.plot = T
)
}
\arguments{
\item{metrics}{A character vector of the desired statistics.  Multiple metrics can be specified.}

\item{title}{User specified plot title. If unspecified, the plot title defaults to the names of the statistics chosen.}

\item{show.intersection}{Show the intersection of 2 performance metrics.}

\item{intersection.only}{Only return the intersection value (cutoff value)}

\item{legend.in.plot}{Put legend within plot area?}

\item{m}{A matrix or data frame with each row being a confusion matrix. The output of 
confusionMatrix(probs.predicted, logicals.expected, cutoff = 'all')}
}
\value{
Returns a ggplot2 line plot
}
\description{
Plots the classification statistics (e.g. true positive rate, true negative rate) as a line plot.
}
