% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPerfCompound.R
\name{plotPerfCompound}
\alias{plotPerfCompound}
\title{Plot compound performance metrics}
\usage{
plotPerfCompound(
  confusion,
  compound.metric = "roc",
  title = NULL,
  show.auc = T,
  auc.rounding = 3,
  show.cutoff.colors = T,
  legend.in.plot = T,
  ...
)
}
\arguments{
\item{confusion}{A matrix or data frame with each row being a confusion matrix. The output of 
confusionMatrix(probs.predicted, logicals.expected, cutoff = 'all')
'weighted' or 'macro'}

\item{...}{Other arguments that can be passed to calcPerfCompound()}

\item{jagged.rough.fix}{Fixed the jagged line when tpr=0 and ppv=0}

\item{metric.names.as.x.y}{If TRUE, will convert colnames: 'tpr' and 'ppv', to: 'x' and 'y'}
}
\value{
Returns a precision-recall matrix compatible for plotting and calculating AUC
}
\description{
Plot compound performance metrics
}
