% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPerformanceRates.R
\name{plotPerformanceRates}
\alias{plotPerformanceRates}
\title{Line plot of classification statistics}
\usage{
plotPerformanceRates(confusion, metrics = c("tpr", "tnr"),
  avg.method = NULL, title = NULL, show.intersection = F,
  intersection.only = F)
}
\arguments{
\item{metrics}{A character vector of the desired statistics.  Multiple metrics can be specified.}

\item{title}{User specified plot title. If unspecified, the plot title defaults to the names of the statistics chosen.}

\item{show.intersection}{Show the intersection of 2 performance metrics.}

\item{intersection.only}{Only return the intersection value (cutoff value)}

\item{m}{A matrix or data frame with each row being a confusion matrix. The output of 
confusionMatrix(probs.predicted, logicals.expected, cutoff = 'all')}
}
\value{
Returns a ggplot2 line plot
}
\description{
Plots the classification statistics (e.g. true positive rate, true negative rate) as a line plot.
}
