% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPR.R
\name{plotPR}
\alias{plotPR}
\title{Plot precision recall}
\usage{
plotPR(confusion, title = NULL, show.auc = T, auc.only = F,
  avg.method = NULL)
}
\arguments{
\item{title}{User specified plot title.}

\item{show.auc}{Shows the area under the under value.}

\item{auc.only}{If TRUE, only returns the AUC value}

\item{m}{A matrix or data frame with each row being a confusion matrix. The output of 
confusionMatrix(probs.predicted, logicals.expected, cutoff = 'all')}
}
\value{
Returns a ggplot2 line plot
}
\description{
Plot precision recall
}
