% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotROC.R
\name{plotROC}
\alias{plotROC}
\title{Plot ROC curve}
\usage{
plotROC(confusion, title = NULL, show.auc = T, auc.only = F,
  avg.method = NULL)
}
\arguments{
\item{title}{User specified plot title.}

\item{show.auc}{Default: TRUE. Shows the AUC on the plot.}

\item{auc.only}{If TRUE, only returns the AUC value}

\item{m}{A matrix or data frame with each row being a confusion matrix. The output of 
confusionMatrix(probs.predicted, logicals.expected, cutoff = 'all')}
}
\value{
Returns a ggplot2 line plot
}
\description{
Plot ROC curve
}
