% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForestCV.R
\name{randomForestCV}
\alias{randomForestCV}
\title{Random forest k-fold cross-validation}
\usage{
randomForestCV(df, colname.response, k = 10, stratify = T, ...)
}
\arguments{
\item{df}{A dataframe of observations as rows and features as columns. \strong{Important:} input matrix/dataframe should
include the response vector as a column.}

\item{colname.response}{The name of a column containing the response classes.}

\item{k}{Number of CV folds.}

\item{stratify}{Stratification ensures that all classes are present in every train/test set when splitting the
data into k-folds (k).}

\item{...}{Other arguments that can be passed to \pkg{randomForest::tuneRF()}, \pkg{randomForest::randomForest()} and
\pkg{mltoolkit::randomForestTrainAndTest()}.}
}
\value{
Returns a list of length(k), with sublists containing the random forest model trained on the train dataset and the
prediction probabilities on the test set.
}
\description{
Performs k-fold cross-validation using random forest as a model, and outputs a model from each fold and the
prediction probabilities. \strong{Important:} input matrix/dataframe should include the response vector as a column.
}
