% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmDonorSuffix.R
\name{rmDonorSuffix}
\alias{rmDonorSuffix}
\title{Remove PCAWG donor name suffix}
\usage{
rmDonorSuffix(chr)
}
\arguments{
\item{chr}{A character vector of PCAWG donor names with the format: PD01004a2}
}
\value{
Removes the a/b/a2/b2 suffix to give PD01004
}
\description{
Remove PCAWG donor name suffix
}
\examples{
rmDonorSuffix(PD01004a2)
}
