% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis_functions.R
\name{runKaplanMeier}
\alias{runKaplanMeier}
\title{Perform Kaplan-Meier Survival Analysis}
\usage{
runKaplanMeier(data)
}
\arguments{
\item{data}{A preprocessed data frame with columns: Time, Event.}
}
\value{
A Kaplan-Meier survival object.
}
\description{
This function fits a Kaplan-Meier survival model to the data and returns the results.
}
\examples{
data <- preprocess_data(generate_employment_data())
km_result <- run_kaplan_meier(data)
}
