% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDataIntoKfolds.R
\name{splitDataIntoKfolds}
\alias{splitDataIntoKfolds}
\title{Split data into k-fold train and test sets}
\usage{
splitDataIntoKfolds(df, k = 10)
}
\arguments{
\item{df}{A dataframe of observations as rows and features as columns}

\item{k}{Number of folds to split the data into}
}
\value{
A list of length(k), containing a sublist of train and test sets
}
\description{
Helper function for createCvTrainTestSets()
}
