% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeFeature.R
\name{standardizeFeature}
\alias{standardizeFeature}
\title{Feature standardization}
\usage{
standardizeFeature(x, na.replace = F, sd.zero.bypass = F, Mean = NULL,
  SD = NULL, Median = NULL)
}
\arguments{
\item{x}{A numeric/integer vector.}

\item{na.replace}{Imputation by median (TRUE or 'median') or mean ('mean').}

\item{sd.zero.bypass}{If standard deviation is zero, return a vector of zeros.}

\item{Mean}{User specified mean value.}

\item{SD}{User specified standard deviation value.}

\item{Median}{User specified median value. Only used for imputation (i.e. if na.replace = T).}
}
\value{
Reurns a vector of standardized values
}
\description{
Standardize a vector with the following transformation: (x - mean)/sd. Standardization makes the values of 
each feature in the data have zero-mean.
}
\details{
Feature scaling is done so that feature values are comparable across features. 

For regularization when fitting a generalized linear model (e.g. with \pkg{glmnet}), feature scaling must be done to 
ensure that regularization penalties are applied with the same magnitude across features. 

In stochastic gradient descent, feature scaling can improve the convergence speed of the algorithm. 

User specified mean and SD can be useful when standardizing train/test sets, where the \emph{test} set is 
standardized using the mean and sd from the \emph{training} set. This is also a helper function for 
standardizeTrainTestSets().
}
\examples{
standardizeFeature(c(1,1,3,5,7,11))
standardizeFeature(c(1,1,0,2,3,2,1,NA), na.replace = 'median')
standardizeFeature(c(1,1,1,1,1), sd.zero.replace = T)
}
