% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeTrainTestSets.R
\name{standardizeTrainTestSets}
\alias{standardizeTrainTestSets}
\title{Standardize train and test datasets together}
\usage{
standardizeTrainTestSets(train, test = NULL, colname.response,
  na.replace = F, sd.zero.bypass = F)
}
\arguments{
\item{train}{Train dataset. A dataframe of observations as rows and features as columns. \strong{Important:} input
matrix/dataframe should include the response vector as a column.}

\item{test}{(Optional) Test dataset. A dataframe of observations as rows and features as columns. \strong{Important:} 
input matrix/dataframe should include the response vector as a column.}

\item{colname.response}{The name of a column containing the response classes.}

\item{na.replace}{Imputation by median (TRUE or 'median') or mean ('mean').}

\item{sd.zero.bypass}{If standard deviation is zero, return a vector of zeros.}
}
\value{
Only train set specified: returns the standardized train set. Both train and test sets specified: returns a list 
containing the standardized train and test sets.
}
\description{
Standardize the train set with the following transformation: (x - mean)/sd. If test set is unspecified, 
returns the standardized train set. If test set is specified, it is standardized using the mean and sd of the 
train set, and a list of the standardized train and test sets are returned.
}
\details{
For more details. See documentation for standardizeFeature().
}
