% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeVector.R
\name{standardizeVector}
\alias{standardizeVector}
\title{Feature standardization}
\usage{
standardizeVector(x, Mean = NULL, SD = NULL, Median = NULL,
  na.replace = F, sd.zero.replace = F)
}
\arguments{
\item{x}{A numeric/integer vector}

\item{Mean}{User specified mean value}

\item{SD}{User specified standard deviation value}

\item{Median}{User specified median value. Only used for imputation (i.e. if na.replace = T)}

\item{na.replace}{Imputation by median (TRUE or 'median') or mean ('mean').}

\item{sd.zero.replace}{If standard deviation is zero, return a vector of zeros}
}
\value{
Reurns a vector of standardized values
}
\description{
Standardize a vector with the following transformation: (x - mean)/sd. Standardization makes the values of each
feature in the data have zero-mean.
}
\details{
This is used for feature scaling so that values are comparable across features. In stochastic gradient descent,
feature scaling can improve the convergence speed of the algorithm.

User specified mean and sd can be useful when standardizing train/test sets, where the \emph{test} set is standardized using
the mean and sd from the \emph{training} set
}
\examples{
standardizeVector(c(1,1,3,5,7,11))
standardizeVector(c(1,1,0,2,3,2,1,NA), na.replace = 'median')
standardizeVector(c(1,1,1,1,1), sd.zero.replace = T)
}
