% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toBinaryResponseMC.R
\name{toBinaryResponseMC}
\alias{toBinaryResponseMC}
\title{Convert multiclass response to all combinations of binary responses}
\usage{
toBinaryResponseMC(chr, as.factor = T)
}
\arguments{
\item{chr}{A character vector or factor}

\item{as.factor}{Option to return a factor instead of character vector}
}
\value{
A data frame of binary responses in relation to each response
}
\description{
Convert multiclass response to all combinations of binary responses
}
\examples{
responses <- c('BRCA1', 'BRCA2', 'none','BRCA1','BRCA1', 'BRCA2', 'none')
toBinaryResponseMC(responses)
}
