% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceClasses.R
\name{balanceClasses}
\alias{balanceClasses}
\title{Balancing of classes by up/down sampling}
\usage{
balanceClasses(df, colnameResponse, scaling = "up")
}
\arguments{
\item{df}{A dataframe of observations as rows and features as columns}

\item{colnameResponse}{The name of a column containing the response classes}

\item{scaling}{If scaling = 'up', less represented classes are sampled up (with replacement) to the highest represented class.
If scaling = 'down', higher represented classes are sampled down to the lowest represented class.}
}
\value{
Returns a dataframe with balanced classes
}
\description{
Data has class imbalance if e.g. there are 100 BRCA deficient donors and 900 BRCA proficient donors.
If scaling = 'up' the 100 BRCA deficient donors are up-sampled to 900. If scaling = 'down' the 900 BRCA deficient donors
are down-sampled to 100. This function also works if there are multiple classes (i.e. response is not binary)
}
\examples{
balanceClasses(df, 'response', scaling = 'up')
}
