% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAUC.R
\name{calcAUC}
\alias{calcAUC}
\title{Calculate AUC}
\usage{
calcAUC(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric/integer vector of discrete x values}

\item{y}{A numeric/integer vector of discrete y values}

\item{na.rm}{Remove NA's?}
}
\value{
Returns the area under the curve
}
\description{
Calculates the area under a curve (AUC) by integration via the trapezoidal rule, where discrete (x,y)
coordinates are given.
}
\examples{
calcAUC(1:10, runif(10))
}
