% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusionMatrix.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{Create confusion matrix(-ces)}
\usage{
confusionMatrix(probs.predicted, logicals.expected, cutoff = "all")
}
\arguments{
\item{probs.predicted}{A numeric vector of prediction probabilities.}

\item{logicals.expected}{A character vector of response labels.}

\item{cutoff}{A single or vector of classification cutoffs ranging from 0 to 1}
}
\value{
A vector or matrix
}
\description{
Create confusion matrix(-ces)
}
