% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusionMatrixMC.R
\name{confusionMatrixMC}
\alias{confusionMatrixMC}
\title{Create confusion matrix(-ces) (multi-class)}
\usage{
confusionMatrixMC(probs.predicted, responses.expected, cutoff = 0.5,
  neg.response = NULL)
}
\arguments{
\item{probs.predicted}{A matrix or data frame containing the prediction probabilities of each class in each column.}

\item{cutoff}{If provided, a horizontal line will be drawn at the specified value to indicate the chosen cutoff.}

\item{neg.response}{The names of the negative response(s). These classes will be excluded from the calculations.}

\item{logicals.expected}{A character vector of response labels. If provided, dots will be coloured by these labels.}
}
\value{
A vector or matrix
}
\description{
Create confusion matrix(-ces) (multi-class)
}
