% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sortedPredictedProbs.R
\name{plot_sortedPredictedProbs}
\alias{plot_sortedPredictedProbs}
\title{Plot sorted prediction probabilities}
\usage{
plot_sortedPredictedProbs(probs.predicted, logicals.expected = NULL,
  cutoff = NULL, show.confusion = T, annotations = NULL, title = NULL)
}
\arguments{
\item{probs.predicted}{A numeric vector of prediction probabilities.}

\item{logicals.expected}{A character vector of response labels. If provided, dots will be coloured by these labels.}

\item{cutoff}{If provided, a horizontal line will be drawn at the specified value to indicate the chosen cutoff.}

\item{show.confusion}{Show confusion matrix?}

\item{annotations}{A vector containining the desired legend names for the responses, e.g. c('BRCA proficient','BRCA deficient')}

\item{title}{User specified plot title.}
}
\value{
A ggplot2 scatter plot
}
\description{
Plot sorted prediction probabilities
}
