% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCvTrainTestSets.R
\name{splitDataIntoKfolds}
\alias{splitDataIntoKfolds}
\title{Split data into k-fold train and test sets}
\usage{
splitDataIntoKfolds(df = NULL, k = 10, return.data = T, show.warnings = T)
}
\arguments{
\item{df}{A dataframe of observations as rows and features as columns}

\item{k}{Number of folds to split the data into}

\item{return.data}{Return the split dataframe or only indexes}

\item{show.warnings}{Show warnings?}
}
\value{
A list of length(k), containing a sublists of train/test data or 
indexes
}
\description{
Helper function for createCvTrainTestSets()
}
