% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformFeatures.R
\name{transformFeatures}
\alias{transformFeatures}
\title{Transform feature columns}
\usage{
transformFeatures(df, formula, colname.response = NULL)
}
\arguments{
\item{df}{A dataframe of features (with or without a response column)}

\item{formula}{A character string describing the mathematical operation, e.g. 'log(x + 1)', where x is the variable to 
transform.}

\item{colname.response}{(Optional) The name of a column containing the response classes. This can be omitted if the 
dataframe only contains feature columns. This is useful if the response vector is a separate object.}
}
\value{
Returns a dataframe of transformed features
}
\description{
Transform the feature columns in a dataframe while excluding the response column (if is exists)
}
\examples{
transformFeatures(df, formula = 'log(x + 1)', colname.response = 'response')
}
